fix(preflight): pin windowsHide on gh/glab auth probe; split oversized preflight modules
- preflight-handler.ts: add windowsHide:true to the auth-status
  execFileAsync probe, matching the other spawn sites in the file, to
  satisfy the Windows console-visibility ratchet.
- Extract command-lookup-spec building (buildCommandLookupSpec(s),
  firstAbsoluteCommandPath, hasAbsoluteCommandPath, and their private
  helpers) into src/relay/relay-command-lookup.ts, re-exported from
  preflight-handler.ts so the public surface (including
  isCommandOnPathForRelay/resolveRelayCommandPath) is unchanged.
- Extract remote/host forge CLI detection probes (detectRemoteAgents,
  detectRemoteForgeClis, isGhAuthenticated, isGlabAuthenticated) from
  agent-detection.ts into agent-forge-detection-probes.ts, re-exporting
  the two still-imported-elsewhere functions.
- Extract local runtime worktree/repo lookup helpers from
  local-preflight-context.ts into local-runtime-worktree-lookup.ts.

All three files are now under the 300-line oxlint max-lines cap.
