+++ src/relay/relay-command-lookup.ts
@@ -1,0 +1,145 @@
+import { userInfo } from 'node:os'
+import path, { win32 } from 'node:path'
+import { buildPosixCommandPathLookupScript } from '../shared/posix-command-path-lookup'
+
+export type CommandLookupSpec = {
+  file: string
+  args: string[]
+  windowsHide?: true
+}
+
+const SUPPORTED_POSIX_SHELLS = new Set(['sh', 'dash', 'bash', 'zsh', 'fish'])
+const CONSERVATIVE_SYSTEM_SHELL_DIRS = new Set(['/bin', '/usr/bin'])
+const AGENT_PATH_PREFIX = '__ORCA_AGENT_PATH__'
+
+export function buildCommandLookupSpec(
+  command: string,
+  platform: NodeJS.Platform,
+  env: NodeJS.ProcessEnv = process.env,
+  accountLoginShell?: string | null
+): CommandLookupSpec {
+  const [spec] = buildCommandLookupSpecs(command, platform, env, accountLoginShell)
+  return spec ?? buildPosixCommandLookupSpec(command, '/bin/sh')
+}
+
+export function buildCommandLookupSpecs(
+  command: string,
+  platform: NodeJS.Platform,
+  env: NodeJS.ProcessEnv = process.env,
+  accountLoginShell?: string | null
+): CommandLookupSpec[] {
+  if (platform === 'win32') {
+    return [{ file: 'where.exe', args: [command], windowsHide: true }]
+  }
+  const trustedShell = pickTrustedPosixShell(
+    env,
+    resolveAccountLoginShell(platform, accountLoginShell)
+  )
+  const specs: CommandLookupSpec[] = []
+
+  if (trustedShell) {
+    specs.push(buildPosixCommandLookupSpec(command, trustedShell))
+  }
+
+  const inheritedPathSpec = buildPosixCommandLookupSpec(command, '/bin/sh')
+  if (!trustedShell || trustedShell !== inheritedPathSpec.file) {
+    specs.push(inheritedPathSpec)
+  }
+
+  return specs
+}
+
+export function hasAbsoluteCommandPath(output: string, platform: NodeJS.Platform): boolean {
+  return firstAbsoluteCommandPath(output, platform) !== null
+}
+
+export function firstAbsoluteCommandPath(output: string, platform: NodeJS.Platform): string | null {
+  const pathOps = platform === 'win32' ? win32 : path
+  for (const raw of output.split(/\r?\n/)) {
+    const line = raw.trim()
+    const resolvedPath =
+      platform === 'win32'
+        ? line
+        : line.startsWith(AGENT_PATH_PREFIX)
+          ? line.slice(AGENT_PATH_PREFIX.length)
+          : ''
+    if (pathOps.isAbsolute(resolvedPath)) {
+      return resolvedPath
+    }
+  }
+  return null
+}
+
+function buildPosixCommandLookupSpec(command: string, shell: string): CommandLookupSpec {
+  const shellName = path.posix.basename(shell).toLowerCase()
+  if (shellName === 'fish') {
+    return { file: shell, args: ['-ilc', buildFishCommandLookupScript(command)] }
+  }
+  return { file: shell, args: [getShellCommandMode(shell), buildShCommandLookupScript(command)] }
+}
+
+function buildShCommandLookupScript(command: string): string {
+  // Why: login shells may define aliases or functions that mask the PATH executable.
+  return [
+    buildPosixCommandPathLookupScript({ kind: 'literal', value: command }),
+    'if [ -n "$resolved" ]; then',
+    `printf '${AGENT_PATH_PREFIX}%s\\n' "$resolved"`,
+    'fi'
+  ].join('\n')
+}
+
+function buildFishCommandLookupScript(command: string): string {
+  const quotedCommand = shellQuote(command)
+  return [
+    `set -l resolved (command -v ${quotedCommand} 2>/dev/null)`,
+    'if test -n "$resolved"',
+    `printf '${AGENT_PATH_PREFIX}%s\\n' "$resolved"`,
+    'end'
+  ].join('\n')
+}
+
+function resolveAccountLoginShell(
+  platform: NodeJS.Platform,
+  accountLoginShell?: string | null
+): string | null {
+  if (accountLoginShell !== undefined) {
+    return accountLoginShell
+  }
+  if (platform === 'win32') {
+    return null
+  }
+  try {
+    return userInfo().shell ?? null
+  } catch {
+    return null
+  }
+}
+
+function pickTrustedPosixShell(
+  env: NodeJS.ProcessEnv,
+  accountLoginShell: string | null
+): string | null {
+  const shell = env.SHELL
+  if (!shell || !path.posix.isAbsolute(shell)) {
+    return null
+  }
+  const shellName = path.posix.basename(shell).toLowerCase()
+  if (!SUPPORTED_POSIX_SHELLS.has(shellName)) {
+    return null
+  }
+  if (accountLoginShell) {
+    return shell === accountLoginShell ? shell : null
+  }
+  return CONSERVATIVE_SYSTEM_SHELL_DIRS.has(path.posix.dirname(shell)) ? shell : null
+}
+
+function getShellCommandMode(shell: string): '-lc' | '-ilc' {
+  const shellName = path.posix.basename(shell).toLowerCase()
+  // Why: bash/zsh/fish users commonly add package-manager bins from interactive
+  // startup files. POSIX sh/dash may not support interactive login flags.
+  return shellName === 'sh' || shellName === 'dash' ? '-lc' : '-ilc'
+}
+
+function shellQuote(value: string): string {
+  return `'${value.replace(/'/g, "'\\''")}'`
+}
